+++ changelog.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 # [4.5.0] - 2018-12-05
+### Added
+ - Tiered menus handle toggling aria-expanded as submenus are opened and closed
+
 ### Changed
  - Updated the toggling behaviour to make the toggleClass optional
  - Reworked SplitDropdown so that it's treated as a single button from keyboard navigation and aria perspectives