+++ changelog.md
@@ -6,9 +6,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-# [4.4.6] - 2018-12-6
+# [4.5.1] - 2018-12-06
 ### Fixed
  - ModalDialog now correctly sets the "aria-modal" attribute
+
+# [4.5.0] - 2018-12-05
+### Added
+ - Tiered menus handle toggling aria-expanded as submenus are opened and closed
+
+### Changed
+ - Updated the toggling behaviour to make the toggleClass optional
+ - Reworked SplitDropdown so that it's treated as a single button from keyboard navigation and aria perspectives
 
 # [4.4.3] - 2018-11-26
 ### Changed